#314 - ServerCache API - change from int to long for ServerCacheStatistics (hitCount, missCount) and ServerCache remove putIfAbsent() method

+++ src/test/java/com/avaje/tests/cache/TestCacheBasic.java
@@ -28,7 +28,7 @@ public class TestCacheBasic extends BaseTestCase {
 
     Country c0 = Ebean.getReference(Country.class, "NZ");
     ServerCacheStatistics statistics = countryCache.getStatistics(false);
-    int hc = statistics.getHitCount();
+    long hc = statistics.getHitCount();
     Assert.assertEquals(1, hc);
     Assert.assertNotNull(c0);
